+++ src/sourcemaps/sourcemaps-wizard.ts
@@ -1,4 +1,4 @@
-// @ts-ignore - clack is ESM and TS complains about that. It works though
+// @ts-expect-error - clack is ESM and TS complains about that. It works though
 import clack from '@clack/prompts';
 import chalk from 'chalk';
 import * as Sentry from '@sentry/node';
@@ -11,6 +11,7 @@ import {
   printWelcome,
   SENTRY_CLI_RC_FILE,
   getOrAskForProjectData,
+  getPackageManager,
 } from '../utils/clack-utils';
 import { isUnicodeSupported } from '../utils/vendor/is-unicorn-supported';
 import type { SourceMapUploadToolConfigurationOptions } from './tools/types';
@@ -28,7 +29,7 @@ import { checkIfMoreSuitableWizardExistsAndAskForRedirect } from './utils/other-
 import { configureAngularSourcemapGenerationFlow } from './tools/angular';
 import type { SupportedTools } from './utils/detect-tool';
 import { detectUsedTool } from './utils/detect-tool';
-import { detectPackageManger } from '../utils/package-manager';
+import { NPM } from '../utils/package-manager';
 import { getIssueStreamUrl } from '../utils/url';
 
 export async function runSourcemapsWizard(
@@ -103,7 +104,7 @@ You can turn this off by running the wizard with the '--disable-telemetry' flag.
     setupCI(selectedTool, authToken, options.comingFrom),
   );
 
-  traceStep('outro', () =>
+  await traceStep('outro', () =>
     printOutro(
       sentryUrl,
       selectedProject.organization.slug,
@@ -312,9 +313,12 @@ SENTRY_AUTH_TOKEN=${authToken}
   }
 }
 
-function printOutro(url: string, orgSlug: string, projectId: string) {
-  const packageManager = detectPackageManger();
-  const buildCommand = packageManager?.buildCommand ?? 'npm run build';
+async function printOutro(
+  url: string,
+  orgSlug: string,
+  projectId: string,
+): Promise<void> {
+  const packageManager = await getPackageManager(NPM);
 
   const issueStreamUrl = getIssueStreamUrl({ url, orgSlug, projectId });
 
@@ -325,7 +329,9 @@ function printOutro(url: string, orgSlug: string, projectId: string) {
    ${chalk.cyan(`Test and validate your setup locally with the following Steps:
 
    1. Build your application in ${chalk.bold('production mode')}.
-      ${chalk.gray(`${arrow} For example, run ${chalk.bold(buildCommand)}.`)}
+      ${chalk.gray(
+        `${arrow} For example, run ${chalk.bold(packageManager.buildCommand)}.`,
+      )}
       ${chalk.gray(
         `${arrow} You should see source map upload logs in your console.`,
       )}