+++ test/react-native/xcode.test.ts
@@ -10,7 +10,7 @@ import {
 } from '../../src/react-native/xcode';
 import chalk from 'chalk';
 import { makeCodeSnippet } from '../../src/utils/clack-utils';
-// @ts-ignore - clack is ESM and TS complains about that. It works though
+// @ts-expect-error - clack is ESM and TS complains about that. It works though
 import * as clack from '@clack/prompts';
 jest.mock('@clack/prompts', () => ({
   log: {